移動中のノックバック処理
かくつき在り

+++ Assets/Scripts/Datas/SpellStatus.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 [CreateAssetMenu]
 public class SpellStatus : ScriptableObject
 {
-    [SerializeField] float damage;
+    [SerializeField] int damage;
     [SerializeField] float pushAmount;
 
-    public float Damage => damage;
+    public int Damage => damage;
     public float PushAmount => pushAmount;
 }
